refactor(benchsdk): split SDK into @benchsdk/api, @benchsdk/worker, and @benchsdk/client umbrella
- Add @benchsdk/api: typed REST client and shared platform types.
- Add @benchsdk/worker: runWorker runtime, BenchmarkReporter, metrics, participants.
- Keep @benchsdk/client as a backwards-compatible umbrella that re-exports from the focused packages and exposes client.runWorker().
- Update pnpm workspace and lockfile for the two new packages.
- Adjust public-api.contract.test.ts to parse the new barrel .d.ts shape with workspace re-exports.
