+++ packages/benchsdk-api/tsup.config.ts
@@ -1,0 +1,16 @@
+import { defineConfig } from 'tsup';
+import pkg from './package.json';
+
+export default defineConfig({
+  entry: {
+    index: 'src/index.ts',
+  },
+  format: ['cjs', 'esm'],
+  dts: true,
+  splitting: false,
+  sourcemap: true,
+  clean: true,
+  define: {
+    __BENCH_VERSION__: JSON.stringify(pkg.version),
+  },
+});
