+++ packages/benchsdk-worker/package.json
@@ -1,9 +1,9 @@
 {
-  "name": "@benchsdk/client",
-  "version": "0.3.0",
+  "name": "@benchsdk/worker",
+  "version": "0.1.0",
   "private": false,
   "type": "module",
-  "description": "Client and worker helpers for the ComputeSDK benchmark orchestrator",
+  "description": "Benchmark worker runtime: claim an assignment, run tasks, flush results and heartbeats",
   "author": "Garrison",
   "license": "MIT",
   "main": "./dist/index.cjs",
@@ -29,20 +29,19 @@
     "pretest": "tsup",
     "test": "vitest run",
     "test:coverage": "vitest run --coverage",
-    "test:integration": "vitest run src/__tests__/integration.test.ts",
     "test:watch": "vitest watch",
     "typecheck": "tsc --noEmit"
   },
   "keywords": [
     "computesdk",
     "benchmark",
-    "observability",
+    "worker",
     "orchestrator"
   ],
   "repository": {
     "type": "git",
     "url": "https://github.com/computesdk/benchmarks.git",
-    "directory": "packages/benchsdk"
+    "directory": "packages/benchsdk-worker"
   },
   "homepage": "https://www.computesdk.com",
   "bugs": {
@@ -52,8 +51,7 @@
     "node": ">=18.0.0"
   },
   "dependencies": {
-    "@benchsdk/api": "workspace:*",
-    "@benchsdk/worker": "workspace:*"
+    "@benchsdk/api": "workspace:*"
   },
   "devDependencies": {
     "@types/node": "^20.0.0",
